refactor(test): separate test project
